Add display progress bar for skinning and unskinning operation

+++ Editor/Core/Components/AssetSkinner/AssetSkinner.cs
@@ -10,10 +10,18 @@ namespace Miniclip.ShapeShifter.Skinner
 {
     public static class AssetSkinner
     {
+        public static void RemoveSkins(string[] assetPaths, bool saveFirst = true)
+        {
+            foreach (string assetPath in assetPaths)
+            {
+                RemoveSkins(assetPath);
+            }
+        }
+
         public static void RemoveSkins(string assetPath)
         {
             string guid = AssetDatabase.AssetPathToGUID(assetPath);
-
+            EditorUtility.DisplayProgressBar("Asset Skinner", $"Removing {assetPath} skin", 0);
             foreach (string game in ShapeShifterConfiguration.Instance.GameNames)
             {
                 string key = ShapeShifterUtils.GenerateUniqueAssetSkinKey(game, guid);
@@ -28,11 +36,22 @@ namespace Miniclip.ShapeShifter.Skinner
                 );
 
                 AssetWatcher.StopWatchingFolder(assetFolder);
-                Directory.Delete(assetFolder, true);
+                EditorUtility.DisplayProgressBar("Asset Skinner", $"Deleting asset skin folder", 0.5f);
+
+                if (Directory.Exists(assetFolder))
+                {
+                    Directory.Delete(assetFolder, true);
+                }
+
+                EditorUtility.DisplayProgressBar("Asset Skinner", $"Staging changes", 0.95f);
+
                 GitUtils.Stage(assetFolder);
             }
 
+            EditorUtility.DisplayProgressBar("Asset Skinner", $"Tracking {assetPath}", 0.95f);
+
             GitUtils.Track(guid);
+            EditorUtility.ClearProgressBar();
         }
 
         internal static void RemoveAllInternalSkins()
@@ -60,13 +79,22 @@ namespace Miniclip.ShapeShifter.Skinner
 
         public static void SkinAsset(string assetPath, bool saveFirst = true)
         {
+            EditorUtility.DisplayProgressBar("Asset Skinner", $"Checking if {assetPath} is supported", 0);
             if (!SupportedTypes.IsSupported(assetPath, out string reason))
             {
+                EditorUtility.ClearProgressBar();
                 return;
             }
-            
+
+            if (IsSkinned(assetPath))
+            {
+                EditorUtility.ClearProgressBar();
+                return;
+            }
+
             if (saveFirst)
             {
+                EditorUtility.DisplayProgressBar("Asset Skinner", $"Saving pending changes", 0.2f);
                 // make sure any pending changes are saved before generating copies
                 ShapeShifterUtils.SavePendingChanges();
             }
@@ -83,15 +111,19 @@ namespace Miniclip.ShapeShifter.Skinner
                     guid
                 );
 
+                EditorUtility.DisplayProgressBar($"Skinning for {game}", $"Checking if its skinned already", 0.3f);
+
                 if (IsSkinned(origin, game))
                 {
                     continue;
                 }
 
+                EditorUtility.DisplayProgressBar($"Skinning for {game}", $"Creating directory for asset", 0.3f);
                 IOUtils.TryCreateDirectory(assetFolder, true);
 
                 string target = Path.Combine(assetFolder, Path.GetFileName(origin));
 
+                EditorUtility.DisplayProgressBar($"Skinning for {game}", $"Copying asset to skin directory", 0.3f);
                 if (AssetDatabase.IsValidFolder(assetPath))
                 {
                     DirectoryInfo targetFolder = Directory.CreateDirectory(target);
@@ -104,16 +136,18 @@ namespace Miniclip.ShapeShifter.Skinner
                     IOUtils.CopyFile(origin + ".meta", target + ".meta");
                 }
 
+                EditorUtility.DisplayProgressBar($"Skinning for {game}", $"Staging changes", 0.3f);
+
                 GitUtils.Stage(assetFolder);
             }
 
+            EditorUtility.DisplayProgressBar($"Asset Skinner", $"Untracking {assetPath}", 0.5f);
             GitUtils.Untrack(guid);
-            
-            GitUtils.Stage(assetPath+".meta");
+            EditorUtility.ClearProgressBar();
         }
 
         public static bool IsSkinned(string assetPath) =>
-            ShapeShifterConfiguration.Instance.GameNames.Any(game => IsSkinned(assetPath, game));
+            ShapeShifterConfiguration.Instance.GameNames.All(game => IsSkinned(assetPath, game));
 
         private static bool IsSkinned(string assetPath, string game)
         {
